Add folder where data is stored

+++ main.py
@@ -9,7 +9,8 @@ def main():
 
 if __name__ == "__main__":
     for file in os.listdir("files"):
-        os.remove(f"files/{file}")
+        if '.csv' in file:
+            os.remove(f"files/{file}")
 
     w = Workouts()
     w.workouts.to_csv("files/workouts.csv")
